feat: reimplement scheduler with Linux 0.11-style dynamic priority
Summary of Changes:

1. Scheduling Algorithm Overhaul:
   - Refactored the core scheduler in `thread.c` from a simple Round-Robin
     to an O(n) complexity algorithm based on Linux 0.11.
   - Introduced dynamic priority adjustments using the `ticks` (counter)
     and `priority` fields.
   - Selection Logic: The scheduler now scans the ready list for the task
     with the highest `ticks` value.

2. Heuristic Priority Boosting:
   - Implemented a global tick redistribution mechanism. When all ready
     tasks exhaust their time slices, the system iterates through ALL
     tasks (including blocked ones) to reset their ticks.
   - Formula: `p->ticks = (p->ticks >> 1) + p->priority`.
   - Rationale: This allows I/O-bound tasks (like the Shell) to accumulate
     ticks while blocked, ensuring they gain higher priority upon waking.
     This effectively provides a smooth, heuristic response for
     interactive processes.
   - If no tasks are ready or have available ticks after redistribution,
     the system explicitly selects the idle thread to maintain CPU
     stability and prevents unnecessary context switching overhead
     during tick resets.

+++ include/magicbox/thread.h
@@ -85,8 +85,8 @@ struct task_struct{
 	pid_t pid;
 	enum task_status status;
 	char name[TASK_NAME_LEN];
-	uint8_t priority;
-	uint8_t ticks;
+	int16_t priority;
+	int16_t ticks;
 	bool is_dyn_link; // 标志当前进程是否是动态链接的，用于优化 swap
 	uint32_t elapsed_ticks;
